feat: Add multi-user profile support with /profile/[id]
- Create dynamic /profile/[id] page with owner detection
- Redirect /profile to /profile/[currentUserId]
- Hide edit/logout buttons for non-owners in ProfileHeader
- Hide Settings tab for non-owners in ProfileNav
- Update API routes to accept ?userId= for viewing others
- Update hooks to pass optional userId parameter
